Fix for Y2038 problem

+++ metrics/metrics_window.go
@@ -236,10 +236,8 @@ func (t *TimeWindow) WindowAverage(now time.Time) float64 {
 // operations to find the index in the bucket list.
 // bucketMutex needs to be held.
 func (t *TimeWindow) timeToIndex(tm time.Time) int {
-	// I don't think this run in 2038 :-)
-	// NB: we need to divide by granularity, since it's a compressing mapping
-	// to buckets.
-	return int(tm.Unix()) / int(t.granularity.Seconds())
+	// Use int64 to avoid Y2038 problem, then safely convert
+	return int(tm.Unix() / int64(t.granularity.Seconds()))
 }
 
 // Record adds a value with an associated time to the correct bucket.
